+++ src/game/ShipVehicleInfo.java
@@ -1,27 +1,27 @@
 package game;
 
 public class ShipVehicleInfo {
-	byte image_index;
-	byte base_cost;
+	int image_index;
+	int base_cost;
 	int max_speed;
-	byte cargo_type;
+	int cargo_type;
 	int capacity;
-	byte running_cost;
-	byte sfx;
-	byte refittable;
+	int running_cost;
+	int sfx;
+	int refittable;
 
 	public ShipVehicleInfo(
 			int i, int j, int k, 
 			int l, int m, int n, int o, int p) 
 	{
-		image_index = (byte) i;
-		base_cost = (byte) j;
+		image_index =  i;
+		base_cost =  j;
 		max_speed = k;
-		cargo_type = (byte) l;
+		cargo_type =  l;
 		capacity = m;
-		running_cost = (byte) n;
-		sfx = (byte) o;
-		refittable = (byte) p;
+		running_cost =  n;
+		sfx =  o;
+		refittable =  p;
 	}
 
 }
